Todos los comandos con procesos sin PIPES (el exit no funciona)
