+++ Makefile
@@ -9,12 +9,12 @@ SRC_FILES = src/main.c \
 			src/parse_utils.c \
 			src/char_utils.c \
 			src/free.c \
-			src/executor.c \
 			src/env/env.c \
 			src/env/env_tools.c \
 			src/build_ins/pwd.c \
 			src/build_ins/echo.c \
 			src/build_ins/cd.c \
+			src/execution/executor.c \
 			src/execution/exec_loop.c \
 			src/execution/exec_commons.c \
