+++ src/execution/exec_loop.c
@@ -12,44 +12,34 @@
 
 #include "minishell.h"
 
+void	exit_error(void)
+{
+	perror("ERROR");
+	exit(EXIT_FAILURE);
+}
+
 int exec_loop(t_backpack *backpack, char **envp)
 {
     int i;
     int j;
+    pid_t	pid_one;
+    int	status;
 
+    status = 0;
     i = backpack->commands_nb;
     j = 1;
     while(j <= i)
     {
-       //printf("||((%s En_Linea %d))||=> %s\n", __FILE__,__LINE__, "ERROR");
-       run_cmd(process_tok(backpack->commands_lst), envp);
+        pid_one = fork();
+        if (pid_one == -1)
+		    exit_error();
+	    else if (pid_one == 0)
+		    executor(backpack, envp);
+       //printf("||((%s En_Linea %d))||=> %s\n", __FILE__,__LINE__, "ERROR"); 
        backpack->commands_lst++; //ESTO PUEDE FALLAR SEGURAMENTE !!!!!!!!!!
        j++;
+       waitpid(pid_one, &status, 0);
     }
     return(1);
 }
 
-//Funcion que genera un array 2D con el comando y sus flags a partir de un tok_lst
-//para poder pasarlo al execv mas adelante
-char **process_tok(tok_lst *token)
-{
-    int i;
-    char **cmd;
-    t_list *act;
-
-    i = ft_lstsize(token->arguments);
-    cmd = (char **) ft_calloc(i + 2, sizeof (char **));
-    if (!cmd)
-        return(NULL);
-    //printf("||((%s En_Linea %d))||=> %s\n", __FILE__,__LINE__, "ERROR");
-    cmd[0] = token->command;
-    act = token->arguments;
-    i = 1;
-    while(act)
-    {
-        cmd[i] = (char *) act->content;
-        i++;
-        act = act->next;
-    }
-    return(cmd);
-}
