+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: frlorenz <frlorenz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/16 19:40:48 by miniore           #+#    #+#             */
-/*   Updated: 2025/06/11 16:34:56 by frlorenz         ###   ########.fr       */
+/*   Updated: 2025/06/11 18:06:24 by frlorenz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,7 +53,7 @@ int main(int argc, char **argv, char **envp)
     (void)argv;
     (void)envp;
     //printf con unas macros que indican el archivo y la linea desde donde se a ejecutado para DEBUGEAR!!!!!!!
-    printf("||((%s En_Linea %d))||=> %s\n", __FILE__, __LINE__, "HOLA ¿¿QUIZAS SOY UN ERROR?? O ¿¿NO??");
+    printf("||((%s En_Linea %d))||=> %s\n", __FILE__, __LINE__, "HOLA ¿¿QUIZAS SOY UN ERROR?? ¿¿O NO??");
     backpack = (t_backpack *)ft_calloc(1, sizeof(t_backpack));     //Mirar si es necesario reservar memoria para backpack
     if(argc != 1)
 		return(EXIT_FAILURE);
@@ -72,7 +72,8 @@ int main(int argc, char **argv, char **envp)
         add_history(input);                 // Arreglar uso del historial si usas ctrl+c en un comando ya usado. Se guarda para la siguiente ejecucion¿?
         if(ft_get_command(backpack, input))
             printf("Syntax error.\n");
-        executor(backpack, envp); // Funcion a la que le pasamos la lista de los comandos y que las cosas se intenten ejecutar.
+        exec_loop(backpack, envp);
+        //executor(backpack, envp); // Funcion a la que le pasamos la lista de los comandos y que las cosas se intenten ejecutar.
         //free_list(comandos, demomento); //MIRAR SEGFAULT CUANDO VARIABLES EXPANDIBLES
         free(input);            //Readline genera malloc para la entrada. En caso de liberarlas no es necesario usar clear_history¿?
     }
